fixed comments from vic
